+++ scripts/winesapos-install.sh
@@ -438,10 +438,18 @@ chown -R 1000.1000 /mnt/home/winesap/Desktop
 echo "Setting up desktop shortcuts complete."
 
 echo "Setting up Mac drivers..."
-# Sound driver.
+# Sound driver for Linux <= 5.12.
 arch-chroot /mnt git clone https://github.com/LukeShortCloud/snd_hda_macbookpro.git -b mac-linux-gaming-stick
 arch-chroot /mnt /bin/zsh snd_hda_macbookpro/install.cirrus.driver.sh
-echo "snd-hda-codec-cirrus" >> /mnt/etc/modules-load.d/winesapos.conf
+echo "snd-hda-codec-cirrus" >> /mnt/etc/modules-load.d/winesapos-sound.conf
+# Sound driver for Linux 5.15.
+# https://github.com/LukeShortCloud/winesapOS/issues/152
+arch-chroot /mnt sh -c 'git clone https://github.com/egorenar/snd-hda-codec-cs8409.git;
+  cd snd-hda-codec-cs8409;
+  export KVER=$(ls -1 /lib/modules/ | grep -P "^5.15");
+  make;
+  make install'
+echo "snd-hda-codec-cs8409" >> /mnt/etc/modules-load.d/winesapos-sound.conf
 # MacBook Pro touchbar driver.
 arch-chroot /mnt ${CMD_YAY_INSTALL} macbook12-spi-driver-dkms
 sed -i s'/MODULES=(/MODULES=(applespi spi_pxa2xx_platform intel_lpss_pci apple_ibridge apple_ib_tb apple_ib_als /'g /mnt/etc/mkinitcpio.conf